fix(host): baked_codec includes signal_graph for NodeType; silence sign-compare + ctor reorder
baked_codec.hpp pulled graph_types.hpp (has NodeId but not NodeType); include
signal_graph.hpp so NodeType resolves. Cast input/output_channels to size_t before the
cap comparison, and reorder the BakedGraphProcessor ctor init list to declaration
order (custom_processors_ before the channel counts).

Skill-Update: skip skill=hosting reason=bake .pulpbake codec — internal host bake plumbing, no user-facing hosting behavior

+++ core/host/src/baked_codec.cpp
@@ -129,8 +129,8 @@ std::optional<BakedPlan> parse_plan_bounded(std::span<const std::uint8_t> bytes)
     plan.input_channels = static_cast<int>(r.u32());
     plan.output_channels = static_cast<int>(r.u32());
     if (!r.ok || plan.input_channels < 0 || plan.output_channels < 0 ||
-        plan.input_channels > kBakedMaxTotalPorts ||
-        plan.output_channels > kBakedMaxTotalPorts) {
+        static_cast<std::size_t>(plan.input_channels) > kBakedMaxTotalPorts ||
+        static_cast<std::size_t>(plan.output_channels) > kBakedMaxTotalPorts) {
         return std::nullopt;
     }
